refactor: view 컨트롤러 분리
