+++ src/main/java/roomescape/ViewController.java
@@ -1,0 +1,18 @@
+package roomescape;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class ViewController {
+
+    @GetMapping(path = {"/", "/admin"})
+    public String wellComePage() {
+        return "admin/index";
+    }
+
+    @GetMapping("/admin/reservation")
+    public String reservationPage() {
+        return "admin/reservation-legacy";
+    }
+}
